64 bit fix for debug builds

+++ code/game/g_main.cpp
@@ -73,16 +73,16 @@ static void ClearAllInUse(void)
 
 void SetInUse(gentity_t *ent)
 {
-	assert(((unsigned int)ent)>=(unsigned int)g_entities);
-	assert(((unsigned int)ent)<=(unsigned int)(g_entities+MAX_GENTITIES-1));
+	assert(((uintptr_t)ent)>=(uintptr_t)g_entities);
+	assert(((uintptr_t)ent)<=(uintptr_t)(g_entities+MAX_GENTITIES-1));
 	unsigned int entNum=ent-g_entities;
 	g_entityInUseBits[entNum/32]|=((unsigned int)1)<<(entNum&0x1f);
 }
 
 void ClearInUse(gentity_t *ent)
 {
-	assert(((unsigned int)ent)>=(unsigned int)g_entities);
-	assert(((unsigned int)ent)<=(unsigned int)(g_entities+MAX_GENTITIES-1));
+	assert(((uintptr_t)ent)>=(uintptr_t)g_entities);
+	assert(((uintptr_t)ent)<=(uintptr_t)(g_entities+MAX_GENTITIES-1));
 	unsigned int entNum=ent-g_entities;
 	g_entityInUseBits[entNum/32]&=~(((unsigned int)1)<<(entNum&0x1f));
 }
